Fix monobank dashboard displaying

+++ src/main/java/com/projects/oleksii/leheza/cashtruck/dto/view/DashboardTransactionDto.java
@@ -11,7 +11,7 @@ public class DashboardTransactionDto {
 
     private String categoryName;
     private String description;
-    private long sum;
+    private double sum;
     private String currencyShortName;
     private long delimiter;
 }
